+++ java-client/src/test/java/org/openapitools/client/api/PortalManagementApiTest.java
@@ -16,7 +16,7 @@ import org.openapitools.client.ApiException;
 import java.io.File;
 import org.openapitools.client.model.OpenConsumerCreateRequestDTO;
 import org.openapitools.client.model.OpenConsumerInfoDTO;
-import org.openapitools.client.model.OpenConsumerSummaryDTO;
+import org.openapitools.client.model.OpenConsumerTokenDTO;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
@@ -415,7 +415,7 @@ public class PortalManagementApiTest {
     public void getConsumerListTest() throws ApiException {
         Integer page = null;
         Integer size = null;
-        List<OpenConsumerSummaryDTO> response = api.getConsumerList(page, size);
+        List<OpenConsumerInfoDTO> response = api.getConsumerList(page, size);
         // TODO: test validations
     }
 
@@ -429,7 +429,7 @@ public class PortalManagementApiTest {
     @Test
     public void getConsumerTokenByAppIdTest() throws ApiException {
         String appId = null;
-        OpenConsumerInfoDTO response = api.getConsumerTokenByAppId(appId);
+        OpenConsumerTokenDTO response = api.getConsumerTokenByAppId(appId);
         // TODO: test validations
     }