+++ spring-boot2/src/main/java/com/apollo/openapi/server/api/PortalManagementApi.java
@@ -7,7 +7,7 @@ package com.apollo.openapi.server.api;
 
 import com.apollo.openapi.server.model.OpenConsumerCreateRequestDTO;
 import com.apollo.openapi.server.model.OpenConsumerInfoDTO;
-import com.apollo.openapi.server.model.OpenConsumerSummaryDTO;
+import com.apollo.openapi.server.model.OpenConsumerTokenDTO;
 import io.swagger.v3.oas.annotations.ExternalDocumentation;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -908,7 +908,7 @@ public interface PortalManagementApi {
         tags = { "Portal Management" },
         responses = {
             @ApiResponse(responseCode = "200", description = "成功获取消费者列表", content = {
-                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = OpenConsumerSummaryDTO.class)))
+                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = OpenConsumerInfoDTO.class)))
             })
         },
         security = {
@@ -920,7 +920,7 @@ public interface PortalManagementApi {
         value = "/openapi/v1/consumers",
         produces = { "application/json" }
     )
-    default ResponseEntity<List<OpenConsumerSummaryDTO>> getConsumerList(
+    default ResponseEntity<List<OpenConsumerInfoDTO>> getConsumerList(
         @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
         @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "10") Integer size
     ) {
@@ -942,7 +942,7 @@ public interface PortalManagementApi {
         tags = { "Portal Management" },
         responses = {
             @ApiResponse(responseCode = "200", description = "成功获取消费者Token", content = {
-                @Content(mediaType = "application/json", schema = @Schema(implementation = OpenConsumerInfoDTO.class))
+                @Content(mediaType = "application/json", schema = @Schema(implementation = OpenConsumerTokenDTO.class))
             })
         },
         security = {
@@ -954,7 +954,7 @@ public interface PortalManagementApi {
         value = "/openapi/v1/consumer-tokens/by-appId",
         produces = { "application/json" }
     )
-    default ResponseEntity<OpenConsumerInfoDTO> getConsumerTokenByAppId(
+    default ResponseEntity<OpenConsumerTokenDTO> getConsumerTokenByAppId(
         @NotNull @Parameter(name = "appId", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "appId", required = true) String appId
     ) {
         return getDelegate().getConsumerTokenByAppId(appId);