+++ spring-boot2/src/main/java/com/apollo/openapi/server/api/PortalManagementApiController.java
@@ -2,7 +2,7 @@ package com.apollo.openapi.server.api;
 
 import com.apollo.openapi.server.model.OpenConsumerCreateRequestDTO;
 import com.apollo.openapi.server.model.OpenConsumerInfoDTO;
-import com.apollo.openapi.server.model.OpenConsumerSummaryDTO;
+import com.apollo.openapi.server.model.OpenConsumerTokenDTO;
 
 
 import org.springframework.beans.factory.annotation.Autowired;